Moved ChatsInProgressAccordion to its own folder
